# Reduce method call memory allocation