--- a/vm/call_frame.go
+++ b/vm/call_frame.go
@@ -261,9 +261,9 @@ func (cfs *callFrameStack) top() callFrame {
 }
 
 func newNormalCallFrame(is *instructionSet, filename string, sourceLine int) *normalCallFrame {
-	return &normalCallFrame{baseFrame: &baseFrame{locals: make([]*Pointer, 15), lPr: 0, fileName: filename, sourceLine: sourceLine}, instructionSet: is, pc: 0}
+	return &normalCallFrame{baseFrame: &baseFrame{locals: make([]*Pointer, 5), lPr: 0, fileName: filename, sourceLine: sourceLine}, instructionSet: is, pc: 0}
 }
 
 func newGoMethodCallFrame(m builtinMethodBody, receiver Object, n, filename string, sourceLine int) *goMethodCallFrame {
-	return &goMethodCallFrame{baseFrame: &baseFrame{locals: make([]*Pointer, 15), lPr: 0, fileName: filename, sourceLine: sourceLine}, method: m, name: n, receiver: receiver}
+	return &goMethodCallFrame{baseFrame: &baseFrame{locals: make([]*Pointer, 5), lPr: 0, fileName: filename, sourceLine: sourceLine}, method: m, name: n, receiver: receiver}
 }