--- a/vm/plugin_integration_test.go
+++ b/vm/plugin_integration_test.go
@@ -1,5 +1,3 @@
-// +build linux
-
 package vm
 
 import (