Benchmark result:

```
benchmarking abcb9c02399812108599eed08aa31311180a726a
benchmarking 8efae0e78c03469eafa684eeb0ff0dfca837885b
x86_64-linux
benchmark                         old ns/op     new ns/op     delta
BenchmarkBasicMath/add-2          9440          8496          -10.00%
BenchmarkBasicMath/subtract-2     8860          8323          -6.06%
BenchmarkBasicMath/multiply-2     8809          8547          -2.97%
BenchmarkBasicMath/divide-2       8878          8693          -2.08%
benchmark                         old allocs     new allocs     delta
BenchmarkBasicMath/add-2          78             75             -3.85%
BenchmarkBasicMath/subtract-2     78             75             -3.85%
BenchmarkBasicMath/multiply-2     78             75             -3.85%
BenchmarkBasicMath/divide-2       78             75             -3.85%
benchmark                         old bytes     new bytes     delta
BenchmarkBasicMath/add-2          3792          3152          -16.88%
BenchmarkBasicMath/subtract-2     3792          3152          -16.88%
BenchmarkBasicMath/multiply-2     3792          3152          -16.88%
BenchmarkBasicMath/divide-2       3792          3152          -16.88%
```